#22 채팅 2개 저장 수정
