+++ src/main/java/com/example/backend/chat/redis/RedisPub.java
@@ -1,7 +1,6 @@
 package com.example.backend.chat.redis;
 
 import com.example.backend.chat.dto.request.ChatMessageRequestDto;
-import com.example.backend.chat.dto.response.ChatMessageResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
