+++ src/main/java/com/example/backend/chat/service/ChatMessageService.java
@@ -34,6 +34,7 @@ public class ChatMessageService {
     private final RedisRepository redisRepository;
     private final ParticipantRepository participantRepository;
     private final RedisPub redisPub;
+    private final ChatMessageService2 chatMessageService2;
 
     @Transactional
     public void sendChatMessage(ChatMessageRequestDto message, String email) {
@@ -56,7 +57,31 @@ public class ChatMessageService {
         // 중간에 ResponseDto 로 변경하는 부분 필요 -> 지금은 LocalDateTime 직렬화 오류 현상 때문에 생략
         log.info("chat.service.ChatMessageService.sendChatMessage().end");
         room.newMessage();
+        this.saveChatMessage(message);
         redisPub.publish(redisRepository.getTopic(room.getRoomId()), message);
+    }
+
+    @Transactional
+    public ChatMessage saveChatMessage(ChatMessageRequestDto message) {
+
+        log.info("chat.service.ChatMessageService.saveChatMessage()");
+        // if 문 안에서 participant 숫자로 read 숫자를 계산
+        Long participantCount = chatMessageService2.getParticipantCount(message.getRoomId());
+        log.info("chat.service.ChatMessageService.saveChatMessage().participantCount = " + participantCount);
+        ChatRoom chatRoom = chatRoomRepository.findById(message.getRoomId()).orElseThrow(
+                () -> new CustomException(ErrorCode.ROOM_NOT_FOUND)
+        );
+        Long notRead = chatRoom.getParticipantList().size() - participantCount;
+        log.info("chat.service.ChatMessageService.saveChatMessage().notRead = " + notRead);
+        if (!Objects.equals(message.getSender(), "[알림]")) {
+            User user = userRepository.findByUsername(message.getSender()).orElseThrow(
+                    () -> new CustomException(ErrorCode.USER_NOT_FOUND)
+            );
+            return chatMessageRepository.save(new ChatMessage(message, user, notRead));
+        }
+        else {
+            return chatMessageRepository.save(new ChatMessage(message));
+        }
     }
 
     // 페이징으로 받아서 무한 스크롤 가능할듯
